feat: Bump rust & dependencies; Use BigDecimal for everything

+++ src-rs/models/type_aliases.rs
@@ -1,3 +1,5 @@
-pub type Time = f64;
-pub type PaxMultiplier = f64;
+use bigdecimal::BigDecimal;
+
+pub type Time = BigDecimal;
+pub type PaxMultiplier = BigDecimal;
 pub type DriverId = String;
